requirements.txt
pymongo>=4.6.0
dnspython>=2.6.1
Faker>=24.0.0
python-dotenv
fastapi
uvicorn
python-multipart